+++ src/test/java/lotto/model/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import lotto.model.LottoNumber;
 import org.assertj.core.api.Assertions;
